Convert tests to use Spock + new testing framework
